requirements.txt
django-import-export
psycopg2-binary
Django==4.2.11